+++ src/whatsapp/routers/group.router.ts
@@ -120,6 +120,16 @@ export class GroupRouter extends RouterBroker {
 
         res.status(HttpStatus.OK).json(response);
       })
+      .get(this.routerPath('acceptInvite'), ...guards, async (req, res) => {
+        const response = await this.inviteCodeValidate<GroupInvite>({
+          request: req,
+          schema: groupInviteSchema,
+          ClassRef: GroupInvite,
+          execute: (instance, data) => groupController.acceptInvite(instance, data),
+        });
+
+        res.status(HttpStatus.OK).json(response);
+      })
       .put(this.routerPath('revokeInviteCode'), ...guards, async (req, res) => {
         const response = await this.groupValidate<GroupJid>({
           request: req,